Idk what to summary this, more in the description
Added SproutConsole.html because its complicated to add it inline in the
py file

Modified SproutConsole.py to load the page properly because it needed to
include socket, uncomment init call and change the port number because
windows won't run it on 5207 for some reason. I left that line in there
commented and changed it to port 80 for testing. I also modified it to
read SproutConsole.html and serve it up properly (it can't be served as
a string, it needs to be a bit-like value (error i got when trying to
run before without encoding)) I also got many errors on missing
parenthesis with print lines so I added parenthesis.

I also modified the text that was the html inside of the
SproutConsole.html so the tabs could be clicked on and have it generate
the new page content in the white space on the right, however, when
clicking on the actual tabs on the left nothing happens. If you call the
functions I created though, they work. Also if we add jquery we can
modify the write function to be jquery based which I am a fan of but it
doesn't have to be done that way.  I was going to try to build out a
Nodes page, but I haven't done that yet.