Event Handlers Galore ;) + <code> block fix + Bootstrap
I fixed the <code> block for the example to display properly, however I
can't get it to be less tabbed in than it is.

I added Bootstrap, which broke the nav buttons, but it fixed the styling
on the <code> block.

I added event handlers to all buttons in the nav bar and the new node +
icon.

I moved most of the html into js variables to allow them to be called as
functions with the handlers an update the nav and content areas. This
will allow setting the active flag on the button in the nav bar for the
page that is active.